pyproject.toml
[project]
name = "telegram-bot"
version = "0.1.0"
description = ""
authors = [
    {name = "Your Name",email = "you@example.com"}
]
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "python-telegram-bot[webhooks] (>=22.5,<23.0)",
    "python-dotenv>=1.0.0",
    "requests>=2.31.0",
    "google-cloud-aiplatform==1.110.0"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry]
package-mode = false
